[Tech] Pools charts requests (#1549)
* add accountLiquidity query

* refactoring queries

* refactoring kensetsu indexer types

* add pool tvl query

* fix price chart isAvailable check
